Fix up PF resource allocation
Before allocating VF resources, allocate the minimum number of TX rings
for the PF.  This works around an issue where some resources may already
be allocated to the PF, blocking VF allocations.

Allocate remaining resources to the PF after assigning to VFs.  This
ensures that the VFs have the maximum possible resources at the expense
of the PF.  This should be confiugrable in the future.

Handle errors when allocating VF resources in a non-fatal manner.  The
driver is attaching to the PF, VF failures shouldn't block that.