Create a structure for the PF to track VF info
HWRM_FUNC_CFG requires that the flags argument always be correct, which
means the PF needs to keep track of it for all uses.  To fulfil this,
add a new child VF info structure in the PF driver to track all VF-specific
details.

Since we have this structure, move the request buffer into here along with
the fw_vf_id.