Fix up VMDq support for VFs
If VFs are enabled, only allocate a single vnic to each function.
This allows targeting a VF with a single L2 context, and since
Intel drivers don't allow VMDq on VFs, shouldn't provide an
incompatability with existing software.

In the future, we will enable VMDq on VFs.

We are essentially mapping VMDq pools to VNICs now.  If there are
no VFs, the PF can use all VNICs as VMDq pools.  If there are VFs,
each function (including the PF) will get a single VNIC.

The Intel restriction of not allowing RSS on VFs without VMDq is
not being emulated at this time.  If it needs to be, we can add that
artificial restriction in the future.