bnxt: fix bnxt_mac_addr_add_op to program VF MAC
mac_addr_add dev_op uses "pools" to indicate VF number.
But the bnxt_mac_addr_add_op doesn't use it correctly.
This patch tries to fix that.

Everytime a MAC is configured for a VF by the PF, it is cached locally.
If more then one MAC addresses is passed, the most recent MAC
will be programmed as the default MAC for the VF.
If the VF intends to change its MAC, the HWRM command will
succeed only if the new MAC matches one of the MACs cached for the VF.

Signed-off-by: Ajit Khaparde <ajit.khaparde@broadcom.com>