Do not waste a filter for the default MAC
When installing MAC filters, do not actually install one that matches
the currently configured default MAC of the VF.  The VF driver will
need to do this regardless, so save on hardware resources.

We still need to track the filter in the list though in case the
default MAC is changed in the future.