Disable functions that call vnic_cfg again.
It appears that calling vnic_cfg triggers problems still.  Disabling
functions until this is resolved.  bd_stall can not be toggled, and
stripq can't be toggled.

At present, the default values for this appear to match the desired
use cases, so this doesn't seem to be an issue.

bd_stall: By default, unserviced VFs will not block traffic to other
functions

stripq: If TX insert is enabled, stripq is enabled.