Map ETH_VMDQ_ACCEPT_HASH_MC to the promiscuous bit.
It's not clear that this is correct, but VFd appears to use it this way.

Also, set ALLMULTI and MCAST when MCAST is set to ensure multicast traffic
is received regardless of the VF driver list.  Because the VF driver
multicast list is in the VFs memory space, we can't read it directly from
the PF driver.