net/bnxt: Allow DPDK to alloc more rings using minimal RM capabilities
When DPDK is interacting with RM aware CHiMP FW, it is allowed to
allocate the desired resources, otherwise ethtool has to be utilized to
first setup resources via L2 driver.

If necessary, calculate Tx and Rx ring count using the new RM API.
Otherwise use the Tx and Rx ring count as-is. Update the cp ring count
based on the Tx and Rx ring count.

Change-Id: Ia7c64eccfdef31b2604d2106aae713e2593cbed0
Reviewed-on: http://gerrit-ccxsw.broadcom.net/59374
Reviewed-by: Ajit Kumar Khaparde <ajit.khaparde@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Randy Schacher <stuart.schacher@broadcom.com>