net/bnxt: fix vlan filtering code path
When hw-vlan-filter is enabled on testpmd, driver should only receive
untagged packets and vlan packets for which the VLAN filter is programmed.
Currently it is receiving all vlan packets. This is because, the default
rule to match on MAC is not getting deleted, when hw-vlan-filter is ON.

This patch fixes the problem, by deleting the default MAC rule and
programming a new rule to receive only untagged packets, when
hw-vlan-filter is enabled & another rule for each vlan, as and when
that vlan is configured on that port.

Bug: CTRL-43091
Signed-off-by: Venkat Duvvuru <venkatkumar.duvvuru@broadcom.com>